% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.R
\name{moe}
\alias{moe}
\title{Calculate margin of error for a proportion or difference in proportions}
\usage{
moe(n, p, q, level = 0.95)
}
\arguments{
\item{n}{How many respondents were there?}

\item{p}{What proportion gave the answer of interest?}

\item{q}{What proportion gave the other answer of interest? (Optional)}

\item{level}{What confidence level should be used? (Default: 0.95)}
}
\value{
An object of class "MOE"; a dataframe giving \code{p}, \code{q},
their \code{Difference}, the margin of error for either the difference or \code{p},
labeled \code{MOE}, the \code{Lower} bound of the confidence interval,
the \code{Upper} bound of the confidence interval,
\code{N} or the number of respondents,
and \code{CIlevel} which is the same as \code{level}.
}
\description{
The margin of error for the proportion of poll respondents giving a
particular answer is \deqn{z\sqrt{\frac{p(1-p)}{n}},}{z * p * (1-p) / n,}
where \code{p} is the proportion of respondents that gave the particular answer,
\code{n} is the sample size of the poll, and \code{z} is the corresponding quantile
of the standard normal distribution for the desired confidence level.
The margin of error for the difference between the proportions of poll
respondents who gave one particular answer versus another one is
\deqn{z\sqrt{\frac{p(1-p) + q(1-q) + 2pq}{n}},}{z * (p*(1-p) + q(1-q) + 2pq) / n,}
where \code{p} is the proportion of respondents that gave one particular answer,
\code{q} is the proportion of respondents who gave the other answer of interest,
\code{n} is the sample size of the poll, and \code{z} is the corresponding quantile
of the standard normal distribution for the desired confidence level.
}
