% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.r
\name{moe}
\alias{moe}
\title{Calculate margin of error for simple probability samples}
\usage{
moe(proportion, n, conf.level = 0.95, digits = 2,
  population.correction = FALSE, population.size = NULL)
}
\arguments{
\item{proportion}{value between 0 and 1 indicating the proportion, such as 0.30 for 30 percent.}

\item{n}{sample size.}

\item{conf.level}{confidence level (defaults to \code{0.95}).}

\item{digits}{number of decimal digits used when formatting the results as APA and human-readable messages (defaults to \code{2}).}

\item{population.correction}{whether or not results should be corrected by population size (defaults to \code{FALSE}).}

\item{population.size}{population size used by the population correction (defaults to \code{NULL}). Only used if \code{population.correction} is set to \code{TRUE}.}
}
\value{
a list with margin of error, confidence level, confidence interval (lower and upper bound), z-value,
results formatted in accordance with APA6 standards, and human-readable interpretation of results.
}
\description{
\code{moe} calculates margin of error and confidence intervals for simple probability
samples, as well as results formatted in accordance with American Psychological Association standards (APA6) and
human-readable interpretations of the results.
}
\section{Assumptions}{

\code{moe} assumes a normal distribution by calculating the z-value from \code{qnorm}.
}

\examples{
\dontrun{
# Margin of error for party with 30\% of voters in a sample of 1,200.
moe(proportion=0.30, n=1200)

# Margin of error for party with 30\% of voters in a sample of 50,000,
# corrected for population size (300,000) and 99\% confidence level.
moe(proportion=0.30, n=50000, conf.level=0.99,
    population.correction=TRUE, population.size=300000)

# Extract confidence interval.
m <- moe(proportion=0.30, n=1200)
m$conf.lower
m$conf.upper

# Extract confidence interval in APA6 format.
m$apa
}
}
\seealso{
\link[stats]{qnorm}
}
